+++ lib/collection/widgets/first_pass_food_item_card.dart
@@ -32,13 +32,12 @@ class FirstPassFoodItemCard extends StatelessWidget {
 
     return Slidable(
       key: Key(foodItem.id),
-      startActionPane: ActionPane(
+      endActionPane: ActionPane(
         motion: const ScrollMotion(),
         children: [
           SlidableAction(
-            // TODO: implement deletion confirmation dialog
             onPressed: (context) => onDeleted!(),
-            backgroundColor: const Color(0xFFFE4A49),
+            backgroundColor: Colors.red,
             foregroundColor: Colors.white,
             icon: Icons.delete,
             label: 'Delete',