+++ lib/collection/widgets/fourth_pass_food_item_card.dart
@@ -22,13 +22,12 @@ class FourthPassFoodItemCard extends StatelessWidget {
   Widget build(BuildContext context) {
     return Slidable(
       key: Key(foodItem.id),
-      startActionPane: ActionPane(
+      endActionPane: ActionPane(
         motion: const ScrollMotion(),
         children: [
           SlidableAction(
-            // TODO: implement deletion confirmation dialog
             onPressed: (context) => onDeleted!(),
-            backgroundColor: const Color(0xFFFE4A49),
+            backgroundColor: Colors.red,
             foregroundColor: Colors.white,
             icon: Icons.delete,
             label: 'Delete',