+++ lib/settings/settings.dart
@@ -1,0 +1,1 @@
+export 'view/settings_help_page.dart';