+++ CHANGELOG.md
@@ -3,8 +3,43 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Feat
+
+- add settings help page with app information
+- **settings**: add source code link to settings help page
+- **settings**: show copyright in settings help page
+- add settings help page with app version
+- update and simplify settings screen
+- update item deletion
+- add delete dialog to fourth pass
+- simplify slidable background red color
+- add long-press-to-delete option to third pass
+- add delete dialog to third pass
+- add long-press-to-delete option to first pass
+- add delete dialog to first pass
+- add long-press-to-delete recipe option
+- standardize delete dialogs
+- add long-press-to-delete collection option
+- change collection form fields to tiles
+- add deletion dialog to collections screen
+- reverse slide to delete direction
+- trim ingredient list and add generating script
+- add script for generating ingredient list
+
+### Fix
+
+- **recipe**: override back button/swipe to save
+- **collection**: override back button/swipe to save
+- **collection**: temporarily remove long-press-to-delete
+- fix wrongly named map key for argument passing for page navigation
+
 ### Refactor
 
+- **collection**: move and rename pop function
+- **collection**: move food item deletion dialog class
+- rename delete recipe to delete recipe dialog
+- remove second ingredients source
+- **recipe**: improve code readability
 - **recipe**: rename bloc private functions
 - **recipe**: standardise ingredients json fields
 - **recipe**: remove unused dependency
@@ -12,10 +47,6 @@ All notable changes to this project will be documented in this file.
 - **recipe**: refactor ingredient loading into bloc and store in state
 - **recipe**: add helper functions to refactor bloc functions
 - **recipe**: use maps instead of lists for arg passing
-
-### Fix
-
-- fix wrongly named map key for argument passing for page navigation
 
 ## 1.1.0 (2022-02-21)